# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "simple description",
        "title": "QrMenuApi",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}/categories/{categoryID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get info of category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.getCategoryOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Update category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "category params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.CategoryInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.updateCategoryOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Delete category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.deleteCategoryOutput"
                        }
                    }
                }
            }
        },
...
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}/categories/{categoryID}/image": {
            "get": {
                "description": "get category image",
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get category image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads category image(png)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Upload image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "categoryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/categories.fileUploadOutput"
                        }
                    }
                }
            }
        },
        "/company/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "returns company info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Get company info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/companies.getCompanyOutput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "categories.deleteCategoryOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "categories.fileUploadOutput": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "categories.getCategoryOutput": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entities.CategoryOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "categories.updateCategoryOutput": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entities.CategoryOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "companies.getCompanyOutput": {
            "type": "object",
            "properties": {
                "company": {
                    "$ref": "#/definitions/entities.CompanyOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "entities.CategoryInput": {
            "type": "object",
            "required": [
                "name",
                "place",
                "visible"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My category name"
                },
                "place": {
                    "type": "integer",
                    "example": 1
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entities.CategoryOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "menu_id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My category name"
                },
                "place": {
                    "type": "integer",
                    "example": 1
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "entities.CompanyOutput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My company name"
                },
                "phone": {
                    "type": "string",
                    "example": "37377777777"
                }
            }
        },
...
    }
}